[csharp] Fix reference error for NUnit 3.2.1
There was an error with the reference to NUnit 3.2.1 assemblies, causing
the reference to be invalid. In Xamarin, the tests fall back to the
bundled NUnit 2.8.7. In Visual Studio, the missing reference causes the
tests to fail because NUnit is not bundled in VS.

This is an incomplete fix. With these changes, tests will fail in
Xamarin because it doesn't support NUnit 3.x. (see
https://forums.xamarin.com/discussion/55863/upgrading-nunit-to-3-0-0-in-my-project-breaks-the-tests).

It looks like the best solution is to roll back to NUnit 2.8.7.